Move unit tests
Add comment to node pool example
